rename vbox/box -> tall/wide
I *think* I have the markup pretty much where I want it. Still need to
verify that alt text is required, but that shouldn't be too hard. I want
to take a look at the tests cause it seems like one or more is writing
to disk which I don't think is a good idea? anyway.