lib: add response_from_file
feature gated behind markup. it's kind of garbage. I don't like the fact
that we can't determine the item kind, potentially there's a file
library that would help? I probably want to move it out of the library
into the server also. I'd need to make that piece of crap good first.

this is just a rough sketch but I can kind of feel it coming together a
bit. v nice.